♻️ share SurrealDbEmbeddedEngine among embedded projects
