+++ SurrealDb.Embedded.Internals/Usings.cs
@@ -1,0 +1,1 @@
+﻿global using CborSerializer = Dahomey.Cbor.Cbor;
