+++ SurrealDb.Embedded.RocksDb/Extensions/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-﻿using SurrealDb.Embedded.RocksDb.Internals;
+﻿using SurrealDb.Embedded.Internals;
 using SurrealDb.Net.Internals;
 
 namespace Microsoft.Extensions.DependencyInjection;
@@ -11,7 +11,7 @@ public static class ServiceCollectionExtensions
 {
     public static SurrealDbBuilder AddRocksDbProvider(this SurrealDbBuilder builder)
     {
-        builder.Services.AddTransient<ISurrealDbRocksDbEngine, SurrealDbRocksDbEngine>();
+        builder.Services.AddTransient<ISurrealDbRocksDbEngine, SurrealDbEmbeddedEngine>();
 
         return builder;
     }
